Split out coap_io code into platform files
