+++ platform/lwip/platform_io.h
@@ -1,0 +1,22 @@
+#include <lwip/udp.h>
+
+/*
+ * This is only included in coap_io.h instead of .c in order to be available for
+ * sizeof in lwippools.h.
+ * Simple carry-over of the incoming pbuf that is later turned into a node.
+ *
+ * Source address data is currently side-banded via ip_current_dest_addr & co
+ * as the packets have limited lifetime anyway.
+ */
+struct coap_packet_t {
+  struct pbuf *pbuf;
+  const coap_endpoint_t *local_interface;
+  uint16_t srcport;
+};
+
+/**
+ * Get the pbuf of a packet. The caller takes over responsibility for freeing
+ * the pbuf.
+ */
+struct pbuf *coap_packet_extract_pbuf(coap_packet_t *packet);
+
