+++ platform/posix/platform_address.c
@@ -1,3 +1,5 @@
+#include "coap_config.h"
+
 #include "address.h"
 
 int
